+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/rest/action/TransportCreateUserAction.java
@@ -49,4 +49,3 @@ public class TransportCreateUserAction extends HandledTransportAction<CreateUser
     }
 
 }
-