+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/rest/request/CreateUserRequest.java
@@ -113,4 +113,3 @@ public class CreateUserRequest extends ActionRequest implements ToXContentObject
     }
 
 }
-